Made a more robust 'complaint' method. See description.
Complaint ignores (for now) reasons of type HttpCodes.TIME_OUT because it triggered too many complaints, which did not cause member eviction because they were not suspect types on the monitor. HttpCode.TIME_OUT complaints could be valid in other scenarios with different considerations than our use case, i.e., time out is not in the scope of our simulation and probably won't ever be.